[AIP-21] Move contrib.operators.mongo_to_s3 providers.amazon.aws.operators.mongo_to_s3
