+++ tests/providers/amazon/aws/operators/test_mongo_to_s3.py
@@ -21,8 +21,8 @@ import unittest
 import mock
 
 from airflow import DAG
-from airflow.contrib.operators.mongo_to_s3 import MongoToS3Operator
 from airflow.models import TaskInstance
+from airflow.providers.amazon.aws.operators.mongo_to_s3 import MongoToS3Operator
 from airflow.utils import timezone
 
 TASK_ID = 'test_mongo_to_s3_operator'
@@ -84,8 +84,8 @@ class TestMongoToS3Operator(unittest.TestCase):
             getattr(self.mock_operator, 'mongo_query')
         )
 
-    @mock.patch('airflow.contrib.operators.mongo_to_s3.MongoHook')
-    @mock.patch('airflow.contrib.operators.mongo_to_s3.S3Hook')
+    @mock.patch('airflow.providers.amazon.aws.operators.mongo_to_s3.MongoHook')
+    @mock.patch('airflow.providers.amazon.aws.operators.mongo_to_s3.S3Hook')
     def test_execute(self, mock_s3_hook, mock_mongo_hook):
         operator = self.mock_operator
